fix(app): separate connectedroot from app wrapper compopnent
Also add in `time` and `h1` to html sanitization allowedTags
