+++ packages/core/components/app/index.js
@@ -2,36 +2,22 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { hot } from 'react-hot-loader/root';
 import { connect } from 'react-redux';
-import RootProviders from 'components/rootProviders';
-import ConnectedRoot from 'components/connectedRoot';
 import ErrorBoundary from 'components/errorBoundary';
-import getRoots from 'selectors/getRoots';
-import getProviders from 'selectors/getProviders';
+import Root from 'components/root';
 import getComponent from 'config/componentMap';
 
 const ErrorMessage = getComponent('error-message');
 const AppContentComponent = getComponent('app');
 
 const App = (props) => {
-  const {
-    error,
-    roots,
-    providers,
-  } = props;
-  const CoreApp = () => (
-    <RootProviders providers={providers}>
-      {roots.map((name) => (
-        <ConnectedRoot key={name} name={name} />
-      ))}
-    </RootProviders>
-  );
+  const { error } = props;
 
   return (
     <ErrorBoundary>
       {error ? (
         <ErrorMessage />
       ) : (
-        <AppContentComponent IrvingApp={CoreApp} />
+        <AppContentComponent IrvingApp={Root} />
       )}
     </ErrorBoundary>
   );
@@ -39,35 +25,26 @@ const App = (props) => {
 
 App.propTypes = {
   /**
-   * Root component configurations
-   */
-  roots: PropTypes.arrayOf(PropTypes.string).isRequired,
-  /**
    * Was there an error loading the page/components?
    */
   error: PropTypes.bool.isRequired,
-  /**
-   * Root provider configurations
-   */
-  providers: PropTypes.arrayOf(PropTypes.object).isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  roots: getRoots(state),
-  providers: getProviders(state),
+const appMapStateToProps = (state) => ({
   error: !! state.error,
 });
 
-const withRedux = connect(mapStateToProps);
+const ConnectedApp = connect(appMapStateToProps)(App);
+
 let hotApp; // eslint-disable-line import/no-mutable-exports
 
 if (
   'production_client' === process.env.IRVING_EXECUTION_CONTEXT ||
   'development_client' === process.env.IRVING_EXECUTION_CONTEXT
 ) {
-  hotApp = hot(withRedux(App));
+  hotApp = hot(ConnectedApp);
 } else {
-  hotApp = withRedux(App);
+  hotApp = ConnectedApp;
 }
 
 /** @component */
