+++ packages/core/config/html.js
@@ -1,9 +1,9 @@
 export const richText = {
   allowedTags: [
-    'h2', 'h3', 'h4', 'h5', 'h6', 'blockquote', 'p', 'a', 'ul', 'ol',
+    'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'blockquote', 'p', 'a', 'ul', 'ol',
     'nl', 'li', 'b', 'i', 'strong', 'em', 'strike', 'code', 'cite', 'hr', 'br',
     'div', 'table', 'thead', 'caption', 'tbody', 'tr', 'th', 'td', 'pre', 'img',
-    'figure', 'figcaption', 'iframe', 'audio', 'video',
+    'figure', 'figcaption', 'iframe', 'audio', 'video', 'time',
   ],
   allowedAttributes: {
     '*': ['class', 'id', 'data-*', 'style'],
